Fix tests and coverage
